First draft of switching over to the WIP page
