+++ app/pathogen/arbovirus/dashboard/(table)/ArboDataTable.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import { columns } from "@/app/pathogen/arbovirus/analyze/columns";
 import { DataTable } from "@/components/ui/data-table/data-table";
 import React, { useContext } from "react";
 import { ArboContext } from "@/contexts/arbo-context";
+import { columns } from "./columns";
 
 //TODO: SeanKennyNF remove this type, the typeguard, and all references to expanding and minimizing visualizations once the redesign is rolled out.
 interface OldArboDataTableProps {
