+++ app/pathogen/arbovirus/dashboard/(table)/columns.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { ColumnDef } from "@tanstack/table-core";
 import React from "react";
 import { Button } from "@/components/ui/button";
 import { ArrowUpDown } from "lucide-react";
