+++ app/pathogen/arbovirus/dashboard/(visualizations)/visualizations-section.tsx
@@ -1,9 +1,9 @@
 import { useCallback } from "react";
 import { useRouter } from "next/navigation";
-import { VisualizationId, VisualizationInformation, addToVisualizationInformation } from "../visualizations/visualizations";
+import { VisualizationId, VisualizationInformation, addToVisualizationInformation } from "../../visualizations/visualizations";
 import { ZoomIn } from "lucide-react";
 import { cn } from '@/lib/utils';
-import { RedesignedArbovirusPageSectionId } from "../../../constants";
+import { RedesignedArbovirusPageSectionId } from "../../../../constants";
 
 export const VisualizationsSection = () => {
     const router = useRouter();
@@ -34,7 +34,7 @@ export const VisualizationsSection = () => {
           <div className="flex py-4">
             <h3 className="w-full text-center text-lg">{visualization.displayName}</h3>
             <button
-              onClick={() => router.push(`visualizations?visualization=${visualization.urlParameter}&referrerRoute=/pathogen/arbovirus/wip%23${RedesignedArbovirusPageSectionId.VISUALIZATIONS}`)}
+              onClick={() => router.push(`visualizations?visualization=${visualization.urlParameter}&referrerRoute=/pathogen/arbovirus/dashboard%23${RedesignedArbovirusPageSectionId.VISUALIZATIONS}`)}
               aria-label="See visualization in fullscreen"
             >
               <ZoomIn />
