+++ app/pathogen/arbovirus/dashboard/page.tsx
@@ -1,116 +1,46 @@
-"use client";
-import React, { useCallback, useContext, useMemo } from "react";
+"use client"
+import { notFound } from "next/navigation";
+import React from "react";
+import { Filters } from "./filters";
+import { VisualizationsSection } from "./(visualizations)/visualizations-section";
+import { ArboDataTable } from "./(table)/ArboDataTable";
+import { ArbovirusMap } from "./(map)/ArbovirusMap";
+import { RedesignedArbovirusPageSectionId } from "../../../constants";
 
-import { CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+export default function RedesignedArbovirusPage() {
+  if(process.env.NEXT_PUBLIC_WEBSITE_REDESIGN_ENABLED !== 'true') {
+    return notFound();
+  }
 
-import { ArbovirusMap } from "@/app/pathogen/arbovirus/dashboard/(map)/ArbovirusMap";
-import { LegendConfiguration, MedianSeroPrevByWHOregion, WHORegionAndArbovirusBar } from "../analyze/recharts";
-import { FilterableField, Filters } from "./filters";
-import { CardConfiguration, CardStyle, CardType, getConfigurationForCard } from "@/components/ui/card-collection/card-collection-types";
-import { CardCollection } from "@/components/ui/card-collection/card-collection";
-import { useMap } from "react-map-gl";
-import { ZoomIn } from "lucide-react";
-import { useRouter } from "next/navigation";
-import { VisualizationId, getVisualizationInformationFromVisualizationId } from "../visualizations/visualizations";
-import { EstimateCountByUnRegionAndArbovirusGraph } from "../analyze/recharts/estimate-count-by-un-region-and-arbovirus-graph";
-import { ArboContext } from "@/contexts/arbo-context";
-import { useArboDataInsights } from "@/hooks/useArboDataInsights";
-
-export default function ArbovirusDashboard() {
-  // Need to make the visualizations dynamic. Unsure how to do this well using CSS.
-  const allMaps = useMap();
-  const router = useRouter();
-  const { filteredData } = useContext(ArboContext);
-  const { getNumberOfUniqueValuesForField } = useArboDataInsights();
-
-  const renderPlotCardContent = useCallback(({cardConfigurations}: {cardConfigurations: CardConfiguration[]}) => {
-    const areLessThanTwoWHORegionsPresentInData = getNumberOfUniqueValuesForField({
-      filteredData: filteredData.filter((dataPoint) => !!dataPoint.whoRegion),
-      fieldName: 'whoRegion'
-    }) < 2;
-
-    const visualizationId = areLessThanTwoWHORegionsPresentInData
-      ? VisualizationId.ESTIMATE_COUNT_BY_UN_REGION_AND_ARBOVIRUS
-      : VisualizationId.ESTIMATE_COUNT_BY_WHO_REGION_AND_ARBOVIRUS;
-
-    const renderVisualization = visualizationId === VisualizationId.ESTIMATE_COUNT_BY_WHO_REGION_AND_ARBOVIRUS 
-      ? () => <WHORegionAndArbovirusBar legendConfiguration={LegendConfiguration.BOTTOM_ALIGNED} />
-      : () => <EstimateCountByUnRegionAndArbovirusGraph legendConfiguration={LegendConfiguration.BOTTOM_ALIGNED} />;
-
-    const visualizationInformation = getVisualizationInformationFromVisualizationId({ visualizationId });
-
-    return (
-      <CardContent className={"px-0 h-full flex flex-col"}>
-        <div className="flex py-4">
-          <h3 className="w-full text-center text-lg">
-            {visualizationInformation.displayName}
-          </h3>
-          <button 
-            onClick={() => router.push(`visualizations?visualization=${getVisualizationInformationFromVisualizationId({ visualizationId }).urlParameter}&referrerRoute=/pathogen/arbovirus/dashboard`)}
-            aria-label="See visualization in fullscreen"
-          >
-            <ZoomIn />
-          </button>
-        </div>
-        {renderVisualization()}
-      </CardContent>
-    );
-  }, [router, filteredData, getNumberOfUniqueValuesForField])
-
-  const renderMapCardContent = useCallback(({cardConfigurations}: {cardConfigurations: CardConfiguration[]}) => (
-    <ArbovirusMap
-      minimizeVisualizations={() => {getConfigurationForCard(cardConfigurations, 'plots', CardType.EXPANDABLE).minimizeCard()}}
-      expandVisualizations={() => {getConfigurationForCard(cardConfigurations, 'plots', CardType.EXPANDABLE).expandCard()}}
-      areVisualizationsExpanded={getConfigurationForCard(cardConfigurations, 'plots', CardType.EXPANDABLE).isExpanded}
-    />
-  ), [])
-
-  const renderFiltersCardContent = useCallback(({ cardConfigurations }: {cardConfigurations: CardConfiguration[]}) => (
-    <>
-      <CardHeader>
-        <CardTitle>Filters</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <Filters excludedFields={[FilterableField.pathogen]} />
-      </CardContent>
-    </>
-  ), [])
-
-  const cardInputData = useMemo(() => [
-    {
-      // order determines where the card will be positioned. Cards with higher order values will be placed to the right of cards with lower order values.
-      order: 1,
-      cardId: 'plots',
-      type: CardType.EXPANDABLE as const,
-      isExpandedByDefault: true,
-      expandedColumnCount: 4,
-      cardClassname: "row-span-2 pr-4 pb-4",
-      renderCardContent: renderPlotCardContent,
-      cardStyle: CardStyle.CARD
-    },
-    {
-      order: 2,
-      cardId: 'map',
-      type: CardType.FILL_REMAINING_SPACE as const,
-      cardClassname: "w-full h-full overflow-hidden row-span-2 relative",
-      renderCardContent: renderMapCardContent,
-      onCardSizeChange: () => {
-        if(allMaps['arboMap']) {
-          allMaps['arboMap'].resize();
-        }
-      },
-      cardStyle: CardStyle.CARD
-    },
-    {
-      order: 3,
-      cardId: 'filters',
-      type: CardType.FIXED as const,
-      columnCount: 2,
-      cardClassname: "row-span-2 overflow-y-auto",
-      renderCardContent: renderFiltersCardContent,
-      cardStyle: CardStyle.CARD
-    }
-  ], [renderPlotCardContent, renderMapCardContent, renderFiltersCardContent, allMaps])
-
-  return <CardCollection cardInputData={cardInputData} columnCountToFill={12} />
+  return (
+    <div className="col-span-12 row-span-2 grid gap-0 grid-cols-12 grid-rows-2 grid-flow-col w-screen overflow-hidden border-box -my-4 -ml-4">
+      <div className="overflow-y-scroll col-span-2 h-full row-span-2 border-black border-r-2">
+        <Filters
+          className="p-4"
+        />
+      </div>
+      <div
+        className={"overflow-y-scroll snap-y scroll-smooth col-span-10 row-span-2"}
+      >
+        <section
+          id={RedesignedArbovirusPageSectionId.MAP}
+          className="w-full h-[95%] snap-start snap-always scroll-smooth overflow-hidden relative row-span-2 border-black border-b-2"
+        >
+          <ArbovirusMap />
+        </section>
+        <section
+          id={RedesignedArbovirusPageSectionId.TABLE}
+          className="w-full h-[95%] snap-start snap-always scroll-smooth overflow-scroll border-black border-b-2"
+        >
+          <ArboDataTable />
+        </section>
+        <section
+          id={RedesignedArbovirusPageSectionId.VISUALIZATIONS}
+          className="w-full h-full snap-start snap-always scroll-smooth grid-cols-2 grid-rows-1 grid overflow-y-scroll gap-y-6 pr-4"
+        >
+          <VisualizationsSection />
+        </section>
+      </div>
+    </div>
+  )
 }
